+++ idp_engine/Theory.py
@@ -312,8 +312,9 @@ class Theory(object):
         # initialize assignments, co_constraints, questions
 
         self.co_constraints, questions = OrderedSet(), OrderedSet()
+        self.constraints = OrderedSet([v.interpret(self)
+                                       for v in self.constraints])
         for c in self.constraints:
-            c.interpret(self)  # may add co_constraints
             c.co_constraints(self.co_constraints)
             # don't collect questions from type constraints
             if not c.is_type_constraint_for:
@@ -321,8 +322,8 @@ class Theory(object):
         for es in self.def_constraints.values():
             for e in es:
                 e.co_constraints(self.co_constraints)
-        for c in self.co_constraints.values():
-            c.interpret(self)
+        self.co_constraints = OrderedSet([c.interpret(self)
+                                          for c in self.co_constraints])
 
         for s in list(questions.values()):
             if s.code not in self.assignments:
@@ -369,7 +370,7 @@ class Theory(object):
             AssertionError: if code is unknown
         """
         _ = self.solver_reified
-        assert any(e.code == code
+        assert any(e.original.code == code
                    for _, e in self.expl_reifs.values()), \
                 f"Cannot enable an unknown law: {code}"
         self.ignored_laws.remove(code)
@@ -387,7 +388,7 @@ class Theory(object):
             AssertionError: if code is unknown
         """
         _ = self.solver_reified
-        assert any(e.code == code
+        assert any(e.original.code == code
                    for _, e in self.expl_reifs.values()), \
                 f"Cannot disable an unknown law: {code}"
         self.ignored_laws.add(code)
@@ -561,7 +562,7 @@ class Theory(object):
             if a.status in [S.CONSEQUENCE, S.ENV_CONSQ, S.UNIVERSAL]:
                 self.assignments.assert__(a.sentence, None, S.UNKNOWN)
         for z3_form, (_, expr) in ps.items():
-            if not (expr and expr.code in self.ignored_laws):
+            if not (expr and expr.original.code in self.ignored_laws):
                 solver.add(z3_form)
 
     def expand(self,
@@ -860,7 +861,7 @@ class Theory(object):
             solver.set("timeout", int(timeout_seconds*1000))
 
         result = solver.check([z3_form for z3_form, (_, expr) in ps.items() if
-                                    not (expr and expr.code in self.ignored_laws)])
+                                    not (expr and expr.original.code in self.ignored_laws)])
 
         solver.set("timeout", int(default_timeout))
         if not timeout_seconds or time.time() - start < timeout_seconds: